When we have more then one server in the writer_hostgroup, we remove all of them to make sure we don't write to the wrong master server.
This is important in split brain situations, where the the management
software may decided to promote both servers to master and proxysql
will start splitting write queryies to multiple master servers,
causing data corruption.

To prevent that, we simply remove all of the servers from the
writer_hostgroup and log a warning.

Signed-off-by: Marian Marinov <mm@yuhu.biz>